RockSlide
RockSlide
When looking at the gate.
RockSlide1 goes on the right and RockSlide 2 goes to the left.
The RockSlide is just the base of mesh for all the smaller rocks of the rockslide.
ModularRock is ment to be scaled and moved on the rock slide and any other place the designers might want to use it.