Updated LockOn Functions
Updated the Lock On mechanic so that the camera transitions between offsets smoothly, also added more camera weight to the enemy so the lock on prioritizes it slightly more. Also included the ability to swap between lock on targets (C & V keys), but I'll have to look into tweaking it some more, because Spherecasting makes switching targets clunky and hard to do because of the limited range.